+++ src/App.tsx
@@ -1,64 +1,12 @@
 import './App.css';
-import { useEffect, useState } from 'react';
-import { FaceLandmarker, FaceLandmarkerOptions, FilesetResolver } from "@mediapipe/tasks-vision";
-import { Color, Euler, Matrix4 } from 'three';
-import { Canvas, useFrame, useGraph } from '@react-three/fiber';
-import { useGLTF } from '@react-three/drei';
+import { useState } from 'react';
+import { Color } from 'three';
+import { Canvas } from '@react-three/fiber';
 import { useDropzone } from 'react-dropzone';
 import CameraPermissions from './camera-permission';
 import ColorSwitcher from './components/ColorSwitcher';
-
-// Global variables for face tracking
-let video: HTMLVideoElement;
-let faceLandmarker: FaceLandmarker;
-let lastVideoTime = -1;
-let blendshapes: any[] = [];
-let rotation: Euler;
-let headMesh: any[] = [];
-
-// Options for Mediapipe FaceLandmarker
-const options: FaceLandmarkerOptions = {
-  baseOptions: {
-    modelAssetPath: `https://storage.googleapis.com/mediapipe-models/face_landmarker/face_landmarker/float16/1/face_landmarker.task`,
-    delegate: "GPU"
-  },
-  numFaces: 1,
-  runningMode: "VIDEO",
-  outputFaceBlendshapes: true,
-  outputFacialTransformationMatrixes: true,
-};
-
-// Avatar component renders the GLTF model and applies blendshapes & head rotation
-function Avatar({ url }: { url: string }) {
-  const { scene } = useGLTF(url);
-  const { nodes } = useGraph(scene);
-
-  useEffect(() => {
-    if (nodes.Wolf3D_Head) headMesh.push(nodes.Wolf3D_Head);
-    if (nodes.Wolf3D_Teeth) headMesh.push(nodes.Wolf3D_Teeth);
-    if (nodes.Wolf3D_Beard) headMesh.push(nodes.Wolf3D_Beard);
-    if (nodes.Wolf3D_Avatar) headMesh.push(nodes.Wolf3D_Avatar);
-    if (nodes.Wolf3D_Head_Custom) headMesh.push(nodes.Wolf3D_Head_Custom);
-  }, [nodes, url]);
-
-  useFrame(() => {
-    if (blendshapes.length > 0) {
-      blendshapes.forEach(element => {
-        headMesh.forEach(mesh => {
-          let index = mesh.morphTargetDictionary[element.categoryName];
-          if (index >= 0) {
-            mesh.morphTargetInfluences[index] = element.score;
-          }
-        });
-      });
-      nodes.Head.rotation.set(rotation.x, rotation.y, rotation.z);
-      nodes.Neck.rotation.set(rotation.x / 5 + 0.3, rotation.y / 5, rotation.z / 5);
-      nodes.Spine2.rotation.set(rotation.x / 10, rotation.y / 10, rotation.z / 10);
-    }
-  });
-
-  return <primitive object={scene} position={[0, -1.75, 3]} />;
-}
+import Avatar from './Avatar';
+import FaceTracking from './FaceTracking';
 
 function App() {
   const [url, setUrl] = useState<string>(
@@ -74,38 +22,19 @@ function App() {
     }
   });
 
-  const setup = async () => {
-    const filesetResolver = await FilesetResolver.forVisionTasks("https://cdn.jsdelivr.net/npm/@mediapipe/tasks-vision@0.10.0/wasm");
-    faceLandmarker = await FaceLandmarker.createFromOptions(filesetResolver, options);
-  };
-
-  const predict = async () => {
-    let nowInMs = Date.now();
-    if (lastVideoTime !== video.currentTime) {
-      lastVideoTime = video.currentTime;
-      const result = faceLandmarker.detectForVideo(video, nowInMs);
-      if (result.faceBlendshapes?.length && result.faceBlendshapes[0].categories) {
-        blendshapes = result.faceBlendshapes[0].categories;
-        const matrix = new Matrix4().fromArray(result.facialTransformationMatrixes![0].data);
-        rotation = new Euler().setFromRotationMatrix(matrix);
-      }
-    }
-    window.requestAnimationFrame(predict);
-  };
-
   const handleStreamReady = (vid: HTMLVideoElement) => {
-    video = vid;
-    video.addEventListener("loadeddata", predict);
+    // Future use if needed, currently FaceTracking manages itself
+    console.log("Video stream ready:", vid);
   };
-
-  useEffect(() => { setup(); }, []);
 
   return (
     <div className="App">
       <CameraPermissions onStreamReady={handleStreamReady} />
+
       <div {...getRootProps({ className: 'dropzone' })}>
         <p>Drag & drop RPM avatar GLB file here</p>
       </div>
+
       <input
         className="url"
         type="text"
@@ -114,7 +43,10 @@ function App() {
           setUrl(`${e.target.value}?morphTargets=ARKit&textureAtlas=1024`)
         }
       />
-      <video className="camera-feed" id="video" autoPlay playsInline></video>
+
+      {/* Moved Mediapipe logic into FaceTracking */}
+      <FaceTracking onStreamReady={handleStreamReady} />
+
       <Canvas style={{ height: 600 }} camera={{ fov: 25 }} shadows>
         <ambientLight intensity={0.5} />
         <pointLight position={[10, 10, 10]} color={new Color(1, 1, 0)} intensity={0.5} castShadow />
@@ -122,9 +54,9 @@ function App() {
         <pointLight position={[0, 0, 10]} intensity={0.5} castShadow />
         <Avatar url={url} />
       </Canvas>
+
       <img className="logo" src="./logo.png" />
-       {/* Add the color switcher here */}
-    <ColorSwitcher />
+      <ColorSwitcher />
     </div>
   );
 }
